Handle multiple roles

+++ backend/routes/api.php
@@ -16,27 +16,31 @@ use App\Http\Controllers\ServiceController;
 use App\Http\Controllers\ServiceOrderController;
 use App\Http\Controllers\ServiceTypeController;
 
-// Rutas de autenticación
+// RUTAS PUBLICAS
 Route::post('login', [AuthController::class, 'login']);
 
 // RUTAS PROTEGIDAS
 Route::middleware('auth:sanctum')->group(function () {
 
-    // USUARIO
-    Route::get('users', [UserController::class, 'index']);
-    Route::put('user/{id}', [UserController::class, 'update'])->middleware('role:Administrador');
-    Route::delete('/user/{id}', [UserController::class, 'destroy'])->middleware('role:Administrador');
-    Route::put('user/{id}/change-password', [UserController::class, 'changePassword'])->middleware('role:Administrador');
+    // AUTH ROUTES
+    Route::post('register', [AuthController::class, 'register'])->middleware('role:Administrador,Jefe Comercial');
+    Route::get('user', [AuthController::class, 'user'])->middleware('role:Administrador,Jefe Comercial,Supervisor,Customer Service,Portero,Romana');
+    Route::post('logout', [AuthController::class, 'logout'])->middleware('role:Administrador,Jefe Comercial,Supervisor,Customer Service,Portero,Romana');
 
-    Route::post('register', [AuthController::class, 'register'])->middleware('role:Administrador');
+
+    // USUARIOS
+    Route::prefix('users')->middleware('role:Administrador,Jefe Comercial')->group(function () {
+        Route::get('/', [UserController::class, 'index']);
+        Route::put('/{id}', [UserController::class, 'update']);
+        Route::delete('/{id}', [UserController::class, 'destroy']);
+        Route::put('/{id}/change-password', [UserController::class, 'changePassword']);
+    });
+
+    // ROLES
     Route::get('roles', [RoleController::class, 'index'])->middleware('role:Administrador,Jefe Comercial');
 
-    // INICIAR O CERRAR SESION
-    Route::get('user', [AuthController::class, 'user']);
-    Route::post('logout', [AuthController::class, 'logout']);
-
     // ATENCION AL CLIENTE 
-    Route::prefix('customers')->group(function () {
+    Route::prefix('customers')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [CustomerController::class, 'index']);
         Route::post('/', [CustomerController::class, 'store']);
         Route::put('/{id}', [CustomerController::class, 'update']);
@@ -45,21 +49,21 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
     // SERVICIOS
-    Route::prefix('services')->group(function () {
-        Route::get('/', [ServiceController::class, 'index']);
-        Route::post('/', [ServiceController::class, 'store']);
-        Route::put('/{id}', [ServiceController::class, 'update']);
-        Route::delete('/{id}', [ServiceController::class, 'destroy']);
-        Route::get('/by-type/{id}', [ServiceController::class, 'getServicesByType']);
+    Route::prefix('services')->middleware('role:Administrador,Jefe Comercial')->group(function () {
+        Route::get('/', [ServiceController::class, 'index'])->middleware('role:Administrador,Jefe Comercial');
+        Route::post('/', [ServiceController::class, 'store'])->middleware('role:Administrador,Jefe Comercial');
+        Route::put('/{id}', [ServiceController::class, 'update'])->middleware('role:Administrador,Jefe Comercial');
+        Route::delete('/{id}', [ServiceController::class, 'destroy'])->middleware('role:Administrador,Jefe Comercial');
+        Route::get('/by-type/{id}', [ServiceController::class, 'getServicesByType'])->middleware('role:Administrador,Jefe Comercial');
     });
 
     // TIPO DE SERVICIOS
-    Route::prefix('services-type')->group(function () {
-        Route::get('/', [ServiceTypeController::class, 'index']);
+    Route::prefix('services-type')->middleware('role:Administrador,Jefe Comercial')->group(function () {
+        Route::get('/', [ServiceTypeController::class, 'index'])->middleware('role:Administrador,Jefe Comercial');
     });
 
     // PRODUCTOS
-    Route::prefix('products')->group(function () {
+    Route::prefix('products')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [ProductController::class, 'index']);
         Route::post('/', [ProductController::class, 'store']);
         Route::put('/{id}', [ProductController::class, 'update']);
@@ -68,7 +72,7 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
     // NEGOCIOS
-    Route::prefix('businesses')->group(function () {
+    Route::prefix('businesses')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [BusinessController::class, 'index']);
         Route::post('/', [BusinessController::class, 'store']);
         Route::put('/{id}', [BusinessController::class, 'update']);
@@ -76,12 +80,12 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
     // RUTAS
-    Route::prefix('routes')->group(function () {
+    Route::prefix('routes')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [RouteController::class, 'index']);
     });
 
     // TARIFAS
-    Route::prefix('rates')->group(function () {
+    Route::prefix('rates')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [RateController::class, 'index']);
         Route::post('/', [RateController::class, 'store']);
         Route::put('/{id}', [RateController::class, 'update']);
@@ -91,8 +95,8 @@ Route::middleware('auth:sanctum')->group(function () {
         Route::put('/updateStatus/{id}', [RateController::class, 'updateStatus']);
     });
 
-    // ordenes de servicio (OS)
-    Route::prefix('orders')->group(function () {
+    // ORDENES DE SERVICIO
+    Route::prefix('orders')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [ServiceOrderController::class, 'index']);
         Route::post('/', [ServiceOrderController::class, 'store']);
         Route::put('/{id}', [ServiceOrderController::class, 'update']);
@@ -108,12 +112,12 @@ Route::middleware('auth:sanctum')->group(function () {
     });
 
     // PLANIFICACION OS
-    Route::prefix('schedule')->group(function () {
+    Route::prefix('schedule')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [ScheduleController::class, 'index']);
     });
 
     // PLANIFICACION
-    Route::prefix('planning')->group(function () {
+    Route::prefix('planning')->middleware('role:Administrador,Jefe Comercial')->group(function () {
         Route::get('/', [PlanningController::class, 'index']);
     });
 });
